Fixes to model nesting; numeric formats

+++ Sources/StadiaMaps/Models/HeightResponse.swift
@@ -17,11 +17,11 @@ public struct HeightResponse: Codable, JSONEncodable, Hashable {
     /** The input polyline. */
     public var encodedPolyline: String?
     /** The list of heights for each point, in meters. Present only if `range` is `false`. Null values indicate missing data. */
-    public var height: [Int]?
+    public var height: [Float]?
     /** The list of ranges and heights for each point in the shape, where each entry is an array of length 2. Present only if `range` is `true`. In each pair, the first element represents the range or distance along the input locations. It is the cumulative distance along the previous coordinates in the shape up to the current coordinate. This value for the first coordinate in the shape will always be 0. The second element in the pair represents the height or elevation at the associated coordinate. The height is null if no height data exists for a given location. Both values are expressed in meters. */
-    public var rangeHeight: [[Int]]?
+    public var rangeHeight: [[Float]]?
 
-    public init(id: String? = nil, shape: [Coordinate]? = nil, encodedPolyline: String? = nil, height: [Int]? = nil, rangeHeight: [[Int]]? = nil) {
+    public init(id: String? = nil, shape: [Coordinate]? = nil, encodedPolyline: String? = nil, height: [Float]? = nil, rangeHeight: [[Float]]? = nil) {
         self.id = id
         self.shape = shape
         self.encodedPolyline = encodedPolyline
